Simplify Record and HLS muxers
