+++ src/main/java/io/antmedia/muxer/HLSMuxer.java
@@ -1,55 +1,34 @@
 package io.antmedia.muxer;
 
-import static org.bytedeco.ffmpeg.global.avcodec.*;
-import static org.bytedeco.ffmpeg.global.avformat.AVFMT_GLOBALHEADER;
+import static org.bytedeco.ffmpeg.global.avcodec.AV_CODEC_ID_AAC;
+import static org.bytedeco.ffmpeg.global.avcodec.AV_CODEC_ID_H264;
+import static org.bytedeco.ffmpeg.global.avcodec.AV_CODEC_ID_H265;
+import static org.bytedeco.ffmpeg.global.avcodec.AV_CODEC_ID_MP3;
+import static org.bytedeco.ffmpeg.global.avcodec.avcodec_parameters_from_context;
 import static org.bytedeco.ffmpeg.global.avformat.AVFMT_NOFILE;
 import static org.bytedeco.ffmpeg.global.avformat.AVIO_FLAG_WRITE;
-import static org.bytedeco.ffmpeg.global.avformat.av_write_frame;
-import static org.bytedeco.ffmpeg.global.avformat.av_write_trailer;
 import static org.bytedeco.ffmpeg.global.avformat.avformat_alloc_output_context2;
-import static org.bytedeco.ffmpeg.global.avformat.avformat_free_context;
-import static org.bytedeco.ffmpeg.global.avformat.avformat_new_stream;
 import static org.bytedeco.ffmpeg.global.avformat.avformat_write_header;
-import static org.bytedeco.ffmpeg.global.avformat.avio_closep;
-import static org.bytedeco.ffmpeg.global.avutil.AVMEDIA_TYPE_AUDIO;
-import static org.bytedeco.ffmpeg.global.avutil.AVMEDIA_TYPE_VIDEO;
-import static org.bytedeco.ffmpeg.global.avutil.AV_PIX_FMT_YUV420P;
-import static org.bytedeco.ffmpeg.global.avutil.AV_ROUND_NEAR_INF;
-import static org.bytedeco.ffmpeg.global.avutil.AV_ROUND_PASS_MINMAX;
 import static org.bytedeco.ffmpeg.global.avutil.av_dict_free;
 import static org.bytedeco.ffmpeg.global.avutil.av_dict_set;
 import static org.bytedeco.ffmpeg.global.avutil.av_rescale_q;
-import static org.bytedeco.ffmpeg.global.avutil.av_rescale_q_rnd;
 import static org.bytedeco.ffmpeg.global.avutil.av_strerror;
 
 import java.io.File;
-import java.io.FilenameFilter;
 import java.io.IOException;
-import java.nio.ByteBuffer;
 import java.nio.file.Files;
 import java.nio.file.StandardOpenOption;
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
 import java.util.Set;
 
-import org.apache.commons.lang3.RandomStringUtils;
-import org.apache.tika.utils.ExceptionUtils;
-import org.bytedeco.ffmpeg.avcodec.AVBSFContext;
-import org.bytedeco.ffmpeg.avcodec.AVBitStreamFilter;
 import org.bytedeco.ffmpeg.avcodec.AVCodec;
 import org.bytedeco.ffmpeg.avcodec.AVCodecContext;
 import org.bytedeco.ffmpeg.avcodec.AVCodecParameters;
 import org.bytedeco.ffmpeg.avcodec.AVPacket;
 import org.bytedeco.ffmpeg.avformat.AVFormatContext;
 import org.bytedeco.ffmpeg.avformat.AVIOContext;
-import org.bytedeco.ffmpeg.avformat.AVStream;
 import org.bytedeco.ffmpeg.avutil.AVDictionary;
 import org.bytedeco.ffmpeg.avutil.AVRational;
-import org.bytedeco.ffmpeg.global.avcodec;
 import org.bytedeco.ffmpeg.global.avformat;
-import org.bytedeco.javacpp.BytePointer;
 import org.red5.server.api.scope.IScope;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -61,8 +40,6 @@ public class HLSMuxer extends Muxer  {
 
 
 	private static final String SEGMENT_SUFFIX_TS = "%04d.ts";
-	private AVBSFContext bsfContext;
-	private long lastDTS = -1;
 
 	protected static Logger logger = LoggerFactory.getLogger(HLSMuxer.class);
 	private String  hlsListSize = "20";
@@ -71,28 +48,15 @@ public class HLSMuxer extends Muxer  {
 
 	private AVRational avRationalTimeBase;
 	private long totalSize;
-	private long partialTotalSize;
 	private long startTime;
 	private long currentTime;
-	private long bitrate;
-	private long bitrateReferenceTime;
 
-	private byte[] extradata = null;
-	protected boolean packetReady = false;
 
-	int videoWidth;
-	int videoHeight;
-	private AVPacket tmpPacket;
 	private boolean deleteFileOnExit = true;
-	private int audioIndex;
-	private int videoIndex;
 	private String hlsFlags;
-	private String streamId;
 
 	private String hlsEncryptionKeyInfoFile = null;
 
-	private Map<Integer, AVRational> codecTimeBaseMap = new HashMap<>();
-	private AVPacket videoPkt;
 	protected StorageClient storageClient = null;
 	private String subFolder = null;
 	private String s3StreamsFolderPath = "streams";
@@ -160,8 +124,6 @@ public class HLSMuxer extends Muxer  {
 				options.put("hls_key_info_file", hlsEncryptionKeyInfoFile);
 			}
 
-			this.bitrate=bitrate;
-
 			logger.info("hls time: {}, hls list size: {} for stream:{}", hlsTime, hlsListSize, streamId);
 
 			int bitrateKbps = bitrate / 1000;
@@ -189,28 +151,14 @@ public class HLSMuxer extends Muxer  {
 			if (this.hlsFlags != null && !this.hlsFlags.isEmpty()) {
 				options.put("hls_flags", this.hlsFlags);
 			}
-			tmpPacket = avcodec.av_packet_alloc();
-			av_init_packet(tmpPacket);
-
-			videoPkt = avcodec.av_packet_alloc();
-			av_init_packet(videoPkt);
 
 			isInitialized = true;
 		}
 
 	}
 
-	public static void writeToFile(String absolutePath, String content) {
-		try {
-			Files.write(new File(absolutePath).toPath(), content.getBytes(), StandardOpenOption.CREATE);
-		} catch (IOException e) {
-			if (logger != null) {
-				logger.error(e.toString());
-			}
-		}
-	}
 
-	private AVFormatContext getOutputFormatContext() {
+	public AVFormatContext getOutputFormatContext() {
 		if (outputFormatContext == null) {
 
 			outputFormatContext= new AVFormatContext(null);
@@ -223,16 +171,9 @@ public class HLSMuxer extends Muxer  {
 		return outputFormatContext;
 	}
 
-	private boolean isCodecSupported(int codecId) {
+	@Override
+	public boolean isCodecSupported(int codecId) {
 		return (codecId == AV_CODEC_ID_H264 || codecId == AV_CODEC_ID_AAC || codecId == AV_CODEC_ID_MP3 || codecId == AV_CODEC_ID_H265);
-	}
-
-	/**
-	 *
-	 * @return the bitrate in last 1 second
-	 */
-	public long getBitrate() {
-		return bitrate;
 	}
 
 	public long getAverageBitrate() {
@@ -246,121 +187,18 @@ public class HLSMuxer extends Muxer  {
 		return 0;
 	}
 
-
-	private  void writePacket(AVPacket pkt, AVRational inputTimebase, AVRational outputTimebase, int codecType)
+	@Override
+	public  void writePacket(AVPacket pkt, AVRational inputTimebase, AVRational outputTimebase, int codecType)
 	{
-		if (outputFormatContext == null || !isRunning.get())  {
-			logger.error("OutputFormatContext is not initialized correctly for {}", file.getName());
-			return;
-		}
-
-
-		long pts = pkt.pts();
-		long dts = pkt.dts();
-		long duration = pkt.duration();
-		long pos = pkt.pos();
-
+		
 		totalSize += pkt.size();
-		partialTotalSize += pkt.size();
-		currentTime = av_rescale_q(dts, inputTimebase, avRationalTimeBase);
+		
+		currentTime = av_rescale_q(pkt.dts(), inputTimebase, avRationalTimeBase);
 		if (startTime == 0) {
 			startTime = currentTime;
-			bitrateReferenceTime = currentTime;
 		}
-
-		if ((currentTime - bitrateReferenceTime) >= 1) {
-			bitrate = partialTotalSize * 8;
-			partialTotalSize = 0;
-			bitrateReferenceTime = currentTime;
-		}
-
-		int ret;
-
-		pkt.duration(av_rescale_q(pkt.duration(), inputTimebase, outputTimebase));
-		pkt.pos(-1);
-
-		if (codecType ==  AVMEDIA_TYPE_VIDEO)
-		{
-			// we don't set startTimeInVideoTimebase here because we only start with key frame and we drop all frames
-			// until the first key frame
-			boolean isKeyFrame = false;
-			if ((pkt.flags() & AV_PKT_FLAG_KEY) == 1) {
-				isKeyFrame = true;
-			}
-
-			ByteBuffer byteBuffer;
-
-			/*
-			 * We add this check because when encoder calls this method the packet needs extra data inside
-			 * However, SFUForwarder calls writeVideoBuffer and the method packets itself there
-			 * To prevent memory issues and crashes we don't repacket if the packet is ready to use from SFU forwarder
-			 */
-			if(!packetReady) {
-				byteBuffer = addExtraDataIfKeyFrame(isKeyFrame, extradata, pkt);
-
-				byteBuffer.position(0);
-
-				//Started to manually packet the frames because we want to add the extra data.
-				tmpPacket.stream_index(pkt.stream_index());
-				tmpPacket.data(new BytePointer(byteBuffer));
-				tmpPacket.size(byteBuffer.limit());
-				tmpPacket.pts(av_rescale_q_rnd(pkt.pts(), inputTimebase, outputTimebase, AV_ROUND_NEAR_INF | AV_ROUND_PASS_MINMAX));
-				tmpPacket.position(pkt.position());
-				tmpPacket.flags(pkt.flags());
-				tmpPacket.dts(av_rescale_q_rnd(pkt.dts(), inputTimebase, outputTimebase, AV_ROUND_NEAR_INF | AV_ROUND_PASS_MINMAX));
-			}
-			else{
-				pkt.pts(av_rescale_q_rnd(pkt.pts(), inputTimebase, outputTimebase, AV_ROUND_NEAR_INF|AV_ROUND_PASS_MINMAX));
-				pkt.dts(av_rescale_q_rnd(pkt.dts(), inputTimebase, outputTimebase, AV_ROUND_NEAR_INF|AV_ROUND_PASS_MINMAX));
-				ret = av_packet_ref(tmpPacket , pkt);
-				if (ret < 0) {
-					logger.error("Cannot copy packet for {}", file.getName());
-					return;
-				}
-			}
-
-			if (bsfContext != null)
-			{
-				ret = av_bsf_send_packet(bsfContext, tmpPacket);
-				if (ret < 0)
-					return;
-
-				while (av_bsf_receive_packet(bsfContext, tmpPacket) == 0)
-				{
-					ret = av_write_frame(outputFormatContext, tmpPacket);
-					if (ret < 0 && logger.isInfoEnabled()) {
-						byte[] data = new byte[64];
-						av_strerror(ret, data, data.length);
-						logger.info("cannot write video frame to muxer. Error: {} stream: {}", new String(data, 0, data.length), file.getName());
-					}
-				}
-			}
-			else {
-				ret = av_write_frame(outputFormatContext, tmpPacket);
-				if (ret < 0 && logger.isInfoEnabled()) {
-					byte[] data = new byte[64];
-					av_strerror(ret, data, data.length);
-					logger.info("cannot write video frame to muxer. Error: {} stream: {}", new String(data, 0, data.length), file.getName());
-				}
-			}
-			av_packet_unref(tmpPacket);
-		}
-		else {
-			pkt.pts(av_rescale_q_rnd(pkt.pts(), inputTimebase, outputTimebase, AV_ROUND_NEAR_INF|AV_ROUND_PASS_MINMAX));
-			pkt.dts(av_rescale_q_rnd(pkt.dts(), inputTimebase, outputTimebase, AV_ROUND_NEAR_INF|AV_ROUND_PASS_MINMAX));
-
-			ret = av_write_frame(outputFormatContext, pkt);
-			if (ret < 0 && logger.isInfoEnabled()) {
-				byte[] data = new byte[64];
-				av_strerror(ret, data, data.length);
-				logger.info("cannot write frame(not video) to muxer. Error is {} stream: {}", new String(data, 0, data.length),  file.getName());
-			}
-		}
-		pkt.pts(pts);
-		pkt.dts(dts);
-		pkt.duration(duration);
-		pkt.pos(pos);
-
+		
+		super.writePacket(pkt, inputTimebase, outputTimebase, codecType);
 	}
 
 
@@ -369,45 +207,13 @@ public class HLSMuxer extends Muxer  {
 	 */
 	@Override
 	public synchronized void writeTrailer() {
-		if (!isRunning.get() || outputFormatContext == null) {
-			logger.warn("HLSMuxer trailer is returning because it's not correct state. Isrunning: {}, outputformatContext: {}", isRunning.get(), outputFormatContext);
-			return;
-		}
-		isRunning.set(false);
+		super.writeTrailer();
 		if (avRationalTimeBase != null) {
 			avRationalTimeBase.close();
 			avRationalTimeBase = null;
 		}
 
-		if (bsfContext != null) {
-			av_bsf_free(bsfContext);
-			bsfContext = null;
-		}
-		if (tmpPacket != null) {
-			av_packet_free(tmpPacket);
-			tmpPacket = null;
-		}
-
-		if (videoPkt != null) {
-			av_packet_free(videoPkt);
-			videoPkt = null;
-		}
-
-		if (audioPkt != null) {
-			av_packet_free(audioPkt);
-			audioPkt = null;
-		}
-
-		av_write_trailer(outputFormatContext);
-
-		/* close output */
-		if ((outputFormatContext.oformat().flags() & AVFMT_NOFILE) == 0)
-			avio_closep(outputFormatContext.pb());
-
-		avformat_free_context(outputFormatContext);
-
-		outputFormatContext = null;
-
+		
 		logger.info("Delete File onexit:{} upload to S3:{} stream:{} hls time:{} hlslist size:{} ", deleteFileOnExit, uploadHLSToS3, streamId, hlsTime, hlsListSize);
 
 		vertx.setTimer(Integer.parseInt(hlsTime) * Integer.parseInt(hlsListSize) * 1000, l -> {
@@ -449,363 +255,28 @@ public class HLSMuxer extends Muxer  {
 		});
 
 
-		isRecording = false;	
-	}
-
-	/**
-	 * {@inheritDoc}
-	 */
-	@Override
-	public synchronized void writePacket(AVPacket pkt, AVCodecContext codecContext) {
-		if (!isRunning.get() || !registeredStreamIndexList.contains(pkt.stream_index()))  {
-			if (time2log % 100 == 0) {
-				logger.warn("not registered stream index {}", file.getName());
-				time2log++;
-			}
-			time2log++;
-			return;
-		}
-		AVStream outStream = outputFormatContext.streams(pkt.stream_index());
-		AVRational codecTimebase = codecTimeBaseMap.get(pkt.stream_index());
-		if(codecContext == null){
-			this.packetReady = true;
-		}
-		else{
-			this.packetReady = false;
-		}
-		writePacket(pkt, codecTimebase,  outStream.time_base(), outStream.codecpar().codec_type());
 	}
 
 
-	@Override
-	public boolean addVideoStream(int width, int height, AVRational videoTimebase, int codecId, int streamIndex,
-			boolean isAVC, AVCodecParameters codecpar) {
-		boolean result = false;
-
-		AVFormatContext outputContext = getOutputFormatContext();
-		if (outputContext != null && isCodecSupported(codecId))
-		{
-			registeredStreamIndexList.add(streamIndex);
-			videoIndex = streamIndex;
-			AVStream outStream = avformat_new_stream(outputContext, null);
-
-			outStream.codecpar().width(width);
-			outStream.codecpar().height(height);
-			outStream.codecpar().codec_id(codecId);
-			outStream.codecpar().codec_type(AVMEDIA_TYPE_VIDEO);
-			outStream.codecpar().format(AV_PIX_FMT_YUV420P);
-			outStream.codecpar().codec_tag(0);
-
-			AVRational timeBase = new AVRational();
-			timeBase.num(1).den(1000);
-			codecTimeBaseMap.put(streamIndex, timeBase);
-			videoWidth = width;
-			videoHeight = height;
-			result = true;
-		}
-		return result;
-
-	}
-
-	/**
-	 * {@inheritDoc}
-	 */
 	@Override
 	public synchronized boolean addStream(AVCodec codec, AVCodecContext codecContext, int streamIndex) {
-
-		AVFormatContext context = getOutputFormatContext();
-
-		//We get parameters because we want to add extra data to the streams to keep the aspect ratio
-		AVCodecParameters codecParameters = new AVCodecParameters();
-		int ret = avcodec_parameters_from_context(codecParameters, codecContext);
-
+		AVCodecParameters codecParameter = new AVCodecParameters();
+		int ret = avcodec_parameters_from_context(codecParameter, codecContext);
 		if (ret < 0) {
 			logger.error("Cannot get codec parameters for {}", streamId);
 		}
-
-		if (context == null) {
-			return false;
-		}
-		if (isCodecSupported(codecContext.codec_id())) {
-			registeredStreamIndexList.add(streamIndex);
-			AVStream outStream = avformat_new_stream(context, codec);
-			outStream.index(streamIndex);
-			if (codecContext.codec_type() == AVMEDIA_TYPE_VIDEO) {
-				videoWidth = codecContext.width();
-				videoHeight = codecContext.height();
-
-				avcodec_parameters_from_context(outStream.codecpar(), codecContext);
-				outStream.time_base(codecContext.time_base());
-				codecTimeBaseMap.put(streamIndex, codecContext.time_base());
-
-				extradata = new byte[codecParameters.extradata_size()];
-
-				if(extradata.length > 0) {
-					BytePointer extraDataPointer = codecParameters.extradata();
-					extraDataPointer.get(extradata).close();
-					extraDataPointer.close();
-					logger.info("extra data 0: {}  1: {}, 2:{}, 3:{}, 4:{}", extradata[0], extradata[1], extradata[2], extradata[3], extradata[4]);
-				}
-			}
-			else {
-
-				ret = avcodec_parameters_from_context(outStream.codecpar(), codecContext);
-				codecTimeBaseMap.put(streamIndex, codecContext.time_base());
-				logger.info("copy codec parameter from context {} stream index: {}", ret,  streamIndex);
-				if (codecContext.codec_type() != AVMEDIA_TYPE_AUDIO) {
-					logger.warn("This should be audio codec for {}", file.getName());
-				}
-				outStream.time_base(codecContext.time_base());
-				codecTimeBaseMap.put(streamIndex, codecContext.time_base());
-			}
-			outStream.codecpar().codec_tag(0);
-
-			if ((context.oformat().flags() & AVFMT_GLOBALHEADER) != 0)
-				codecContext.flags( codecContext.flags() | AV_CODEC_FLAG_GLOBAL_HEADER);
-
-		}
-		return true;
+		
+		//call super directly because no need to add bit stream filter 
+		return super.addStream(codecParameter, codecContext.time_base(), streamIndex);
 	}
 
-
 	@Override
-	public boolean addStream(AVCodecParameters codecParameters, AVRational timebase, int streamIndex) 
+	public synchronized boolean addStream(AVCodecParameters codecParameters, AVRational timebase, int streamIndex) 
 	{
-		boolean result = false;
-		AVFormatContext outputContext = getOutputFormatContext();
-		logger.debug("Codec type = {} - Codec tag = {} - Codec id = {} ", codecParameters.codec_type(),codecParameters.codec_tag(), codecParameters.codec_id());
-		if (outputContext != null && isCodecSupported(codecParameters.codec_id()))
-		{
-			AVStream outStream = avformat_new_stream(outputContext, null);
-			if (codecParameters.codec_type() == AVMEDIA_TYPE_VIDEO)
-			{
-				videoIndex = outStream.index();
-				AVBitStreamFilter h264bsfc = av_bsf_get_by_name("h264_mp4toannexb");
-				bsfContext = new AVBSFContext(null);
-
-				extradata = new byte[codecParameters.extradata_size()];
-
-				if(extradata.length > 0) {
-					BytePointer extraDataPointer = codecParameters.extradata();
-					extraDataPointer.get(extradata).close();
-					extraDataPointer.close();
-					logger.info("extra data 0: {}  1: {}, 2:{}, 3:{}, 4:{}", extradata[0], extradata[1], extradata[2], extradata[3], extradata[4]);
-				}
-
-				int ret = av_bsf_alloc(h264bsfc, bsfContext);
-				if (ret < 0) {
-					logger.info("cannot allocate bsf context for {}", file.getName());
-					return false;
-				}
-
-				ret = avcodec_parameters_copy(bsfContext.par_in(), codecParameters);
-				if (ret < 0) {
-					logger.info("cannot copy input codec parameters for {}", file.getName());
-					return false;
-				}
-				bsfContext.time_base_in(timebase);
-
-				ret = av_bsf_init(bsfContext);
-				if (ret < 0) {
-					logger.info("cannot init bit stream filter context for {}", file.getName());
-					return false;
-				}
-
-				ret = avcodec_parameters_copy(outStream.codecpar(), bsfContext.par_out());
-				if (ret < 0) {
-					logger.info("cannot copy codec parameters to output for {}", file.getName());
-					return false;
-				}
-				videoWidth = outStream.codecpar().width();
-				videoHeight = outStream.codecpar().height();
-				outStream.time_base(bsfContext.time_base_out());
-				outStream.codecpar().codec_tag(0);
-			}
-			else {
-				audioIndex = outStream.index();
-				int ret = avcodec_parameters_copy(outStream.codecpar(), codecParameters);
-				if (ret < 0) {
-					logger.info("Cannot get codec parameters for {}", file.getName());
-					return false;
-				}
-
-				if (codecParameters.codec_type() != AVMEDIA_TYPE_AUDIO) {
-					logger.warn("codec type should be audio but it is {}" , codecParameters.codec_type());
-
-				}
-				outStream.codecpar().codec_tag(0);
-			}
-
-			outStream.time_base(timebase);
-			codecTimeBaseMap.put(outStream.index(), timebase);
-			registeredStreamIndexList.add(streamIndex);
-			result = true;
-		}
-		return result;
+		bsfVideoName = "h264_mp4toannexb";
+		return super.addStream(codecParameters, timebase, streamIndex);
 	}
-
-	/**
-	 * {@inheritDoc}
-	 */
-	@Override
-	public synchronized boolean prepareIO() {
-		AVFormatContext context = getOutputFormatContext();
-
-		/**
-		 * We need to extract addedStream information in some cases because we treat audio and video separate
-		 * In addStream for example, if we don't check this we end up removing the muxer completely if one of the operations fail.
-		 */
-		if (isRunning.get()) {
-			logger.warn("HLS Muxer is already running for stream: {} so it's not preparing io again and returning", streamId);
-			return false;
-		}
-
-		int ret = 0;
-
-		if ((context.oformat().flags() & AVFMT_NOFILE) == 0) {
-			AVIOContext pb = new AVIOContext(null);
-
-			ret = avformat.avio_open(pb,  file.getAbsolutePath(), AVIO_FLAG_WRITE);
-			if (ret < 0) {
-				logger.warn("Could not open output file: {} ", file.getAbsolutePath());
-				return false;
-			}
-			context.pb(pb);
-		}
-
-		AVDictionary optionsDictionary = null;
-
-		if (!options.isEmpty()) {
-			optionsDictionary = new AVDictionary();
-			Set<String> keySet = options.keySet();
-			for (String key : keySet) {
-				av_dict_set(optionsDictionary, key, options.get(key), 0);
-			}
-
-		}		
-
-		ret = avformat_write_header(context, optionsDictionary);		
-		if (ret < 0 && logger.isWarnEnabled()) {
-			byte[] data = new byte[1024];
-			av_strerror(ret, data, data.length);
-			logger.warn("could not write header. File: {} Error: {}", file.getAbsolutePath(), new String(data, 0, data.length));
-			return false;
-		}
-
-		if (optionsDictionary != null) {
-			av_dict_free(optionsDictionary);
-		}
-		isRunning.set(true);
-		return true;
-	}
-
-	/**
-	 * {@inheritDoc}
-	 */
-	@Override
-	public synchronized void writePacket(AVPacket avpacket, AVStream inStream) {
-		if (!isRunning.get() || !registeredStreamIndexList.contains(avpacket.stream_index()))  {
-			if (time2log % 100 == 0) {
-				logger.warn("not registered stream index {}", file.getName());
-				time2log = 0;
-			}
-			time2log++;
-			return;
-		}
-		int streamIndex;
-		if (inStream.codecpar().codec_type() == AVMEDIA_TYPE_VIDEO) {
-			streamIndex = videoIndex;
-		}
-		else if (inStream.codecpar().codec_type() == AVMEDIA_TYPE_AUDIO) {
-			streamIndex = audioIndex;
-		}
-		else {
-			logger.error("Undefined codec type for {} stream index: {}", file.getName(), inStream.index());
-			return;
-		}
-
-		AVStream outStream = getOutputFormatContext().streams(streamIndex);
-		int index = avpacket.stream_index();
-		avpacket.stream_index(streamIndex);
-		this.packetReady = true;
-		writePacket(avpacket, inStream.time_base(),  outStream.time_base(), outStream.codecpar().codec_type());
-		avpacket.stream_index(index);
-
-	}
-
-	@Override
-	public void writeAudioBuffer(ByteBuffer audioFrame, int streamIndex, long timestamp) {
-		if (!isRunning.get()) {
-			if (time2log  % 100 == 0) {
-				logger.warn("Not writing AudioBuffer for {} because Is running:{}", file.getName(), isRunning.get());
-				time2log = 0;
-			}
-			time2log++;
-			return;
-		}
-
-		audioPkt.stream_index(streamIndex);
-		audioPkt.pts(timestamp);
-		audioPkt.dts(timestamp);
-		audioFrame.rewind();
-		audioPkt.flags(audioPkt.flags() | AV_PKT_FLAG_KEY);
-		audioPkt.data(new BytePointer(audioFrame));
-		audioPkt.size(audioFrame.limit());
-		audioPkt.position(0);
-
-		writePacket(audioPkt, (AVCodecContext)null);
-
-		av_packet_unref(audioPkt);
-
-	}
-
-	@Override
-	public void writeVideoBuffer(ByteBuffer encodedVideoFrame, long dts, int frameRotation, int streamIndex,
-								 boolean isKeyFrame,long firstFrameTimeStamp, long pts) {
-		/*
-		 * this control is necessary to prevent server from a native crash
-		 * in case of initiation and preparation takes long.
-		 * because native objects like videoPkt can not be initiated yet
-		 */
-		if (!isRunning.get()) {
-			if (time2log % 100 == 0) {
-				logger.warn("Not writing to VideoBuffer for {} because Is running:{}", file.getName(), isRunning.get());
-				time2log = 0;
-			}
-			time2log++;
-			return;
-		}
-
-		videoPkt.stream_index(streamIndex);
-		videoPkt.pts(pts);
-		videoPkt.dts(dts);
-
-		encodedVideoFrame.rewind();
-		if (isKeyFrame) {
-			videoPkt.flags(videoPkt.flags() | AV_PKT_FLAG_KEY);
-		}
-
-		BytePointer bytePointer = new BytePointer(encodedVideoFrame);
-		videoPkt.data(bytePointer);
-		videoPkt.size(encodedVideoFrame.limit());
-		videoPkt.position(0);
-
-
-		writePacket(videoPkt, (AVCodecContext)null);
-
-		av_packet_unref(videoPkt);
-	}
-
-
-
-	public int getVideoWidth() {
-		return videoWidth;
-	}
-
-	public int getVideoHeight() {
-		return videoHeight;
-	}
-
+	
 	public String getHlsListSize() {
 		return hlsListSize;
 	}
